requirements.txt
# Ultra-Advanced Portfolio Optimization System Requirements
# ========================================================

# Core Python packages
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Quantum Computing
qiskit>=0.39.0
qiskit-optimization>=0.4.0
qiskit-aer>=0.11.0
qiskit-ibmq-provider>=0.19.0
dwave-ocean-sdk>=6.0.0
dimod>=0.12.0

# Deep Learning and AI
torch>=1.12.0
torchvision>=0.13.0
torch-geometric>=2.2.0
transformers>=4.21.0
stable-baselines3>=1.6.0
optuna>=3.0.0
ray[tune]>=2.0.0
mlflow>=1.28.0

# Financial Data and APIs
yfinance>=0.1.87
alpha-vantage>=2.3.1
quandl>=3.7.0
FRED>=3.1

# Sentiment Analysis and NLP
textblob>=0.17.1
tweepy>=4.14.0
newspaper3k>=0.2.8
vaderSentiment>=3.3.2

# Web Framework and APIs
fastapi>=0.85.0
uvicorn>=0.18.0
websockets>=10.4
requests>=2.28.0
httpx>=0.23.0

# Database and Storage
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
redis>=4.3.0
influxdb>=5.3.0
psutil>=5.9.0

# Data Processing and Analysis
plotly>=5.10.0
seaborn>=0.11.0
matplotlib>=3.5.0
ta>=0.10.2
sklearn>=0.0

# Configuration and Utilities
PyYAML>=6.0
python-dotenv>=0.20.0
click>=8.1.0
rich>=12.5.0
tqdm>=4.64.0

# Async and Concurrency
aiohttp>=3.8.0
aiofiles>=0.8.0
asyncio-throttle>=1.0.2

# Security and Authentication
cryptography>=37.0.0
PyJWT>=2.4.0
passlib>=1.7.4
bcrypt>=3.2.0

# Monitoring and Logging
prometheus-client>=0.14.0
statsd>=4.0.1
sentry-sdk>=1.9.0

# Testing
pytest>=7.1.0
pytest-asyncio>=0.19.0
pytest-cov>=3.0.0
factory-boy>=3.2.0

# Development Tools
black>=22.6.0
flake8>=5.0.0
mypy>=0.971
isort>=5.10.0
pre-commit>=2.20.0

# Scheduling and Task Management
schedule>=1.2.0
celery>=5.2.0

# Mathematical Optimization
cvxpy>=1.2.0
pulp>=2.6.0
gurobipy>=9.5.0  # Commercial license required

# Additional Financial Libraries
TA-Lib>=0.4.25  # Technical Analysis Library
bt>=0.2.9  # Backtesting framework
zipline-reloaded>=2.2.0  # Backtesting engine
empyrical>=0.5.5  # Financial risk metrics
ffn>=0.3.6  # Financial functions

# Network and Communication
zmq>=0.0.0
paramiko>=2.11.0

# Data Validation
pydantic>=1.10.0
marshmallow>=3.17.0

# Time Series Analysis
statsmodels>=0.13.0
arch>=5.3.0  # GARCH models

# Caching and Performance
joblib>=1.1.0
numba>=0.56.0

# Environment and Configuration
python-decouple>=3.6

# Optional: GPU acceleration (if CUDA available)
# torch-audio>=0.12.0
# torch-text>=0.13.0

# Optional: Additional quantum simulators
# cirq>=1.0.0
# pennylane>=0.25.0

# Optional: Additional ML frameworks
# tensorflow>=2.9.0
# keras>=2.9.0
# xgboost>=1.6.0
# lightgbm>=3.3.0

# Development and Jupyter
jupyter>=1.0.0
jupyterlab>=3.4.0
ipywidgets>=8.0.0

# Documentation
sphinx>=5.1.0
sphinx-rtd-theme>=1.0.0

